feat(tipping): show message preview + last activity in tips list
Reuse the send flow's ChatSummary -> ConversationReference mapping for the
Tips chat list instead of leaving preview/lastActivity unpopulated.

- Extract ChatSummary.toConversationReference() into chat-ui, moving
  formatPreview out of ContactListBuilder (text/cash previews)
- Add lastActivity to ConversationReference so rows can sort/show recency
- ContactListBuilder and TipFlowViewModel now both call the shared mapper
- Tip rows render real message previews and last-activity timestamps
