+++ apps/flipcash/features/tipping/build.gradle.kts
@@ -16,4 +16,5 @@ dependencies {
     implementation(project(":apps:flipcash:shared:chat-ui"))
     implementation(project(":apps:flipcash:shared:shareable"))
     implementation(project(":apps:flipcash:shared:tipping"))
+    implementation(project(":apps:flipcash:shared:tokens"))
 }
