+++ apps/flipcash/features/tipping/src/main/kotlin/com/flipcash/app/tipping/internal/screens/TipsScreen.kt
@@ -122,7 +122,7 @@ private fun TipChatRow(
             )
 
             ChatRowTrailing(
-                lastActivity = null,
+                lastActivity = chat.lastActivity,
                 unreadCount = chat.unreadCount,
                 canOpen = true,
             )
